Changed local schedule length and index to be uint16_t to allow for local schedules longer than 512.

+++ src/gttcan.c
@@ -116,7 +116,7 @@ void gttcan_process_frame(gttcan_t *gttcan, uint32_t can_frame_id, uint64_t data
 
 void gttcan_get_local_schedule(gttcan_t *gttcan, global_schedule_ptr_t global_schedule_ptr)
 {
-    uint8_t local_schedule_index = 0;
+    uint16_t local_schedule_index = 0;
     for (int i = 0; i < gttcan->global_schedule_length; i++)
     {
         if (global_schedule_ptr[i].node_id == gttcan->node_id)
